moved lollipop.ts to types dir
