+++ apps/io-session-manager-internal/src/types/lollipop.ts
